backend/server/requirements.txt
flask==3.0.3
python-dotenv==1.0.1
rank-bm25==0.2.2
chromadb==0.5.5        
sentence-transformers==2.7.0  
tqdm==4.66.4